+++ tests/testthat/test_wdpa_clean.R
@@ -17,8 +17,11 @@ test_that("single country with eez", {
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
   # fetch data
-  x <- wdpa_clean(suppressWarnings(wdpa_fetch(
-    "MHL", wait = TRUE, verbose = TRUE)))
+  x <- wdpa_clean(
+    suppressWarnings(
+      wdpa_fetch("MHL", wait = TRUE, verbose = TRUE, check_version = FALSE)
+    )
+  )
   # run tests
   expect_is(x, "sf")
   expect_true(all(names(x) %in% wdpa_column_names))
@@ -37,8 +40,9 @@ test_that("single country without eez", {
   skip_on_github_workflow("Windows")
   # fetch data
   x <- wdpa_clean(
-    suppressWarnings(wdpa_fetch("LIE", wait = TRUE)),
-    verbose = TRUE, geometry_precision = 1000)
+    suppressWarnings(wdpa_fetch("LIE", wait = TRUE, check_version = FALSE)),
+    verbose = TRUE, geometry_precision = 1000
+  )
   # run tests
   expect_gt(nrow(x), 0)
   expect_true(all(wdpa_column_names %in% names(x)))
@@ -55,8 +59,10 @@ test_that("single country with simplification", {
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
   # fetch data
-  x <- wdpa_clean(suppressWarnings(wdpa_fetch("MHL", wait = TRUE)),
-                  simplify_tolerance = 100)
+  x <- wdpa_clean(
+    suppressWarnings(wdpa_fetch("MHL", wait = TRUE, check_version = FALSE)),
+    simplify_tolerance = 100
+  )
   # run tests
   expect_is(x, "sf")
   expect_true(all(wdpa_column_names %in% names(x)))
@@ -73,8 +79,10 @@ test_that("single country without overlap removal", {
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
   # fetch data
-  x <- wdpa_clean(suppressWarnings(wdpa_fetch("BDI", wait = TRUE)),
-                  erase_overlaps = FALSE)
+  x <- wdpa_clean(
+    suppressWarnings(wdpa_fetch("BDI", wait = TRUE, check_version = FALSE)),
+    erase_overlaps = FALSE
+  )
   # run tests
   expect_gt(nrow(x), 0)
   expect_true(all(wdpa_column_names %in% names(x)))
@@ -91,7 +99,7 @@ test_that("country with MULTIPOINT protected areas", {
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
   # fetch data
-  x <- wdpa_fetch("BOL", wait = TRUE)
+  x <- wdpa_fetch("BOL", wait = TRUE, check_version = FALSE)
   x_points <- vapply(x$geometry, inherits, logical(1),
                      c("POINT", "MULTIPOINT"))
   x <- x[c(which(x_points), 15), , drop = FALSE]
@@ -105,7 +113,7 @@ test_that("country with MULTIPOLYGON protected area", {
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
   # fetch data
-  x <- wdpa_fetch("BOL", wait = TRUE)
+  x <- wdpa_fetch("BOL", wait = TRUE, check_version = FALSE)
   y <- suppressWarnings(wdpa_clean(x, erase_overlaps = FALSE,
                                    geometry_precision = 10000))
   p1 <- x[x$WDPAID == 555592679, ]
@@ -120,7 +128,7 @@ test_that("country with super invalid MULTIPOLYGON data", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
-  x <- wdpa_fetch("GAB", wait = TRUE)
+  x <- wdpa_fetch("GAB", wait = TRUE, check_version = FALSE)
   y <- suppressWarnings(wdpa_clean(x, erase_overlaps = TRUE))
   expect_is(y, "sf")
   expect_true(all(y$STATUS %in% default_retain_status))
@@ -130,7 +138,7 @@ test_that("geometries in non-geometry column", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
-  x <- wdpa_fetch("GAB", wait = TRUE)
+  x <- wdpa_fetch("GAB", wait = TRUE, check_version = FALSE)
   geom_col <- attr(x, "sf_column")
   attr(x, "sf_column") <- "shape"
   names(x)[names(x) == geom_col] <- "shape"
@@ -144,8 +152,9 @@ test_that("single country with no valid non-empty geometries", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
-  x <-
-    suppressWarnings(wdpa_clean(wdpa_fetch("SOM", wait = TRUE)))
+  x <- wdpa_clean(
+    suppressWarnings(wdpa_fetch("SOM", wait = TRUE, check_version = FALSE))
+  )
   y <- wdpa_clean(wdpa_fetch("MHL", wait = TRUE))
   crs <- paste("+proj=cea +lon_0=0 +lat_ts=30 +x_0=0",
     "+y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs")
@@ -167,8 +176,9 @@ test_that("retain UNESCO Biosphere reserves", {
   skip_on_github_workflow("Windows")
   # fetch data
   x <- wdpa_clean(
-    suppressWarnings(wdpa_fetch("MHL", wait = TRUE)),
-    exclude_unesco = FALSE)
+    suppressWarnings(wdpa_fetch("MHL", wait = TRUE, check_version = FALSE)),
+    exclude_unesco = FALSE
+  )
   # run tests
   expect_is(x, "sf")
   expect_true(all(names(x) %in% wdpa_column_names))
@@ -187,7 +197,7 @@ test_that("protected areas that need prepr", {
   skip_if_not_installed("prepr")
   skip_on_github_workflow("Windows")
   # fetch data
-  x <- suppressWarnings(wdpa_fetch("USA", wait = TRUE))
+  x <- suppressWarnings(wdpa_fetch("USA", wait = TRUE, check_version = FALSE))
   x <- x[x$MARINE == "2", , drop = FALSE]
   y <- wdpa_clean(x, erase_overlaps = FALSE)
   # run tests
@@ -201,8 +211,9 @@ test_that("custom retain_status", {
   skip_on_github_workflow("Windows")
   # fetch data
   x <- wdpa_clean(
-    suppressWarnings(wdpa_fetch("MHL", wait = TRUE)),
-    retain_status = c("Designated", "Proposed"))
+    suppressWarnings(wdpa_fetch("MHL", wait = TRUE, check_version = FALSE)),
+    retain_status = c("Designated", "Proposed")
+  )
   # run tests
   expect_is(x, "sf")
   expect_true(all(names(x) %in% wdpa_column_names))
@@ -220,7 +231,7 @@ test_that("NULL retain_status", {
   skip_if_not(curl::has_internet())
   skip_on_github_workflow("Windows")
   # fetch data
-  x <- suppressWarnings(wdpa_fetch("MHL", wait = TRUE))
+  x <- suppressWarnings(wdpa_fetch("MHL", wait = TRUE, check_version = FALSE))
   y <- wdpa_clean(x, retain_status = NULL, erase_overlaps = TRUE)
   # run tests
   expect_is(y, "sf")